Break uptest & crddiff into two separate binaries in preparation of moving uptest under to the Crossplane organization
