+++ cmd/uptest/main.go
@@ -16,27 +16,21 @@
 package main
 
 import (
-	"log"
 	"os"
 	"path/filepath"
 	"strings"
-	"syscall"
 
 	"gopkg.in/alecthomas/kingpin.v2"
 
 	"github.com/upbound/uptest/internal"
 	"github.com/upbound/uptest/internal/config"
-	"github.com/upbound/uptest/internal/crdschema"
 )
 
 var (
 	app = kingpin.New("uptest", "Automated Test Tool for Upbound Official Providers").DefaultEnvars()
-	// e2e command
+	// e2e command (single command is preserved for backward compatibility)
+	// and we may have further commands in the future.
 	e2e = app.Command("e2e", "Run e2e tests for manifests by applying them to a control plane and waiting until a given condition is met.")
-	// crddiff command and sub-commands
-	cmdCRDDiff  = app.Command("crddiff", "A tool for checking breaking API changes between two CRD OpenAPI v3 schemas. The schemas can come from either two revisions of a CRD, or from the versions declared in a single CRD.")
-	cmdRevision = cmdCRDDiff.Command("revision", "Compare the first schema available in a base CRD against the first schema from a revision CRD")
-	cmdSelf     = cmdCRDDiff.Command("self", "Use OpenAPI v3 schemas from a single CRD")
 )
 
 var (
@@ -59,26 +53,9 @@ var (
 	onlyCleanUptestResources = e2e.Flag("only-clean-uptest-resources", "While deletion step, only clean resources that were created by uptest").Default("false").Bool()
 )
 
-var (
-	revisionDiffOptions = getCRDdiffCommonOptions(cmdRevision)
-	selfDiffOptions     = getCRDdiffCommonOptions(cmdSelf)
-)
-
-func getCRDdiffCommonOptions(cmd *kingpin.CmdClause) *crdschema.CommonOptions {
-	opts := &crdschema.CommonOptions{}
-	cmd.Flag("enable-upjet-extensions", "Enables diff extensions for the CRDs generated by upjet. "+
-		"An example extension is the processing of the x-kubernetes-validations CEL rules generated by upjet.").Default("false").BoolVar(&opts.EnableUpjetExtensions)
-	return opts
-}
-
 func main() {
-	switch kingpin.MustParse(app.Parse(os.Args[1:])) {
-	case e2e.FullCommand():
+	if kingpin.MustParse(app.Parse(os.Args[1:])) == e2e.FullCommand() {
 		e2eTests()
-	case cmdRevision.FullCommand():
-		crdDiffRevision()
-	case cmdSelf.FullCommand():
-		crdDiffSelf()
 	}
 }
 
@@ -128,44 +105,4 @@ func e2eTests() {
 	}
 
 	kingpin.FatalIfError(internal.RunTest(o), "cannot run e2e tests successfully")
-}
-
-var (
-	baseCRDPath     = cmdRevision.Arg("base", "The manifest file path of the CRD to be used as the base").Required().ExistingFile()
-	revisionCRDPath = cmdRevision.Arg("revision", "The manifest file path of the CRD to be used as a revision to the base").Required().ExistingFile()
-)
-
-func crdDiffRevision() {
-	crdDiff, err := crdschema.NewRevisionDiff(*baseCRDPath, *revisionCRDPath, crdschema.WithRevisionDiffCommonOptions(revisionDiffOptions))
-	kingpin.FatalIfError(err, "Failed to load CRDs")
-	reportDiff(crdDiff)
-}
-
-var (
-	crdPath = cmdSelf.Arg("crd", "The manifest file path of the CRD whose versions are to be checked for breaking changes").Required().ExistingFile()
-)
-
-func crdDiffSelf() {
-	crdDiff, err := crdschema.NewSelfDiff(*crdPath, crdschema.WithSelfDiffCommonOptions(selfDiffOptions))
-	kingpin.FatalIfError(err, "Failed to load CRDs")
-	reportDiff(crdDiff)
-}
-
-func reportDiff(crdDiff crdschema.SchemaCheck) {
-	versionMap, err := crdDiff.GetBreakingChanges()
-	kingpin.FatalIfError(err, "Failed to compute CRD breaking API changes")
-
-	l := log.New(os.Stderr, "", 0)
-	breakingDetected := false
-	for v, d := range versionMap {
-		if d.Empty() {
-			continue
-		}
-		breakingDetected = true
-		l.Printf("Version %q:\n", v)
-		l.Println(crdschema.GetDiffReport(d))
-	}
-	if breakingDetected {
-		syscall.Exit(1)
-	}
 }
